requirements.txt
pandas
matplotlib
seaborn
google-genai
groq
jinja2
python-dotenv
streamlit
openpyxl
pyarrow
html2text
weasyprint
cerebras-cloud-sdk

